+++ packages/dartantic_ai/test/test_helpers/run_provider_test.dart
@@ -71,7 +71,15 @@ const providerTestCaps = <String, Set<ProviderTestCaps>>{
     ProviderTestCaps.thinking,
     ProviderTestCaps.mediaGeneration,
   },
-  'mistral': {ProviderTestCaps.chat, ProviderTestCaps.embeddings},
+  'mistral': {
+    ProviderTestCaps.chat,
+    ProviderTestCaps.embeddings,
+    // TODO: Re-enable multiToolCalls when mistralai_dart fixes streaming bug
+    // https://github.com/davidmigloz/langchain_dart/issues/913
+    // When fixed: 1) Update mistralai_dart dependency to fixed version
+    //             2) Add ProviderTestCaps.multiToolCalls back to this set
+    //             3) Run tool calling tests to verify: dart test test/tool_calling_test.dart -n "mistral"
+  },
   'cohere': {
     ProviderTestCaps.chat,
     ProviderTestCaps.embeddings,